+++ feature/home/src/main/java/com/teamoffroad/feature/home/data/remote/service/EmblemService.kt
@@ -1,7 +1,7 @@
 package com.teamoffroad.feature.home.data.remote.service
 
 import com.teamoffroad.core.common.data.remote.response.BaseResponse
-import com.teamoffroad.feature.home.data.remote.response.emblem.EmblemsResponseDto
+import com.teamoffroad.feature.home.data.remote.response.EmblemsResponseDto
 import com.teamoffroad.feature.home.data.remote.response.UserQuestsResponseDto
 import retrofit2.http.GET
 import retrofit2.http.Header
